requirements.txt
albumentations==0.0.13
appnope==0.1.0
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.24
chardet==3.0.4
cloudpickle==0.5.5
cycler==0.10.0
Cython==0.28.5
dask==0.19.0
decorator==4.3.0
entrypoints==0.2.3
html5lib==1.0.1
idna==2.7
imgaug==0.2.6
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.4.7.1
kiwisolver==1.0.1
Markdown==3.0
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
munch==2.3.2
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numpy==1.15.1
opencv-python==3.4.2.17
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pretrainedmodels==0.7.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.1.0
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-slugify==1.2.5
pytz==2018.5
PyWavelets==1.0.0
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorboardX==1.4
tensorflow==0.12.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
